Chest scanning!!
Added proper treasure chest scanning for spells, materials, items, and
enemies.

Added an option to display "simple names" (without the ID) instead of
both the name and the ID.

Added notification if a spell is new (does not appear on the Progress
screen).

Add Panel:on_player_spawned.

Change how image fallbacks work: it's now part of the Panel functions.